requirements.txt
beautifulsoup4==4.14.3
fastapi==0.135.1
langchain==1.2.10
langchain_community==0.4.1
langchain_core==1.2.17
langchain_google_genai==4.2.1
pydantic==2.12.5
python-dotenv==1.2.2
